Refactor tree tokenizers into separate classes.
Main reason is improvement in code quality: we were essentially switching
on an enum for polymorphic behavior (bad, bad).

Also: add an additional folder to separate the tokenizers into two
groups---ngram and tree---since they are logically different.

Add a convenience "tokenizers/tokenizers.h" which allows for inclusion of
all tokenizers.