Add support for "opaque typedefs".
These should be used in place of any ordinary typedef (or type alias via
"using") when possible---this ensures that each unique typedef actually
defines a new type, as opposed to being transparent (allowing different
identifiers to collide).

For example, the following would be legal with regular typedefs/aliases:

    doc_id d_id{ 1 };
    term_id t_id;
    t_id = d_id;

but with the new MAKE_IDENTIFIER or MAKE_NUMERIC_IDENTIFIER, the above
code will no longer compile.

This change *shouldn't* introduce any runtime penalty, as the type
checking is done as part of the compilation process, and there are no
virtual functions.