Fix performance regression in classifiers from forward_index refactoring.
The fix involves a few parts: first, search_primary() now returns a
shared_ptr to the postings data---this allows us to avoid having to
allocate memory in the caller thread for an object that is (likely) to
still be in the cache, while still being robust to its potential removal
from the cache (the shared_ptr guarantees the caller that the object will
be live while it is being used).

Second, and perhaps more importantly, the classifiers no longer use
counts() on the forward_index, but instead use search_primary() directly,
and then call counts() on the resulting postings_data object. This
eliminates copying of the map of counts for every iteration that visits a
document. For small datasets, this will bring performance back to nearly
what it was when the entire dataset was required to be in memory.

This requires an API change: search_primary() of disk_index is now public
so the classifiers may use it---I think the 25x performance benefit
outweighs the downsides here.