Refactor disk_index to use CRTP instead of inheritance.
This is motivated by the fact that there was really only one virtual
method being overridden in derived classes (except for search_primary
being overridden in cached_index), and that there was code duplication
between the two in the index creation process. Now things are a little
more clean, at the cost of one little traits class to get around typedef
issues.