Massive refactoring of disk_index and subclasses.
Introduces the p_impl idiom to allow us to change disk_index internals
without breaking binary compatibility with client code.

We still need to introduce this idiom to the derived class indexes as
well, which will help tremendously in reducing the amount of recompiling
that goes on when private index internals change.