Fix denominator in average precision calculation.
It was broken before in that it used the cutoff point as the denominator
instead of considering the total number of relevant documents. The fix
from the pull request is correct, but only if the cutoff point was not
specified to be something smaller than the number of relevant documents.